feat: add dev.env, docker-compose for dev env setup
