+++ backend-flask/config.py
@@ -1,0 +1,1 @@
+## TODO: Define configurations for a MySQL environment
