+++ backend-flask/dev.env
@@ -1,0 +1,4 @@
+MYSQL_USER="campgen-user"
+MYSQL_PASSWORD="campgen-pass"
+MYSQL_ROOT_PASSWORD="IamRoot69"
+MYSQL_DATABASE="campgen_db"
