+++ backend-flask/docker-compose.yaml
@@ -1,0 +1,27 @@
+version: '3.3'
+services:
+  db:
+    image: mysql:5.7
+    restart: always
+    env_file:
+      - dev.env
+    environment:
+      MYSQL_DATABASE: "${MYSQL_DATABASE}"
+      # Additional user
+      MYSQL_USER: '${MYSQL_USER}'
+      # You can use whatever password you like
+      MYSQL_PASSWORD: '${MYSQL_PASSWORD}'
+      # Password for root access
+      MYSQL_ROOT_PASSWORD: '${MYSQL_ROOT_PASSWORD}'
+    ports:
+      # <Port exposed> : <MySQL Port running inside container>
+      - '3306:3306'
+    expose:
+      # Opens port 3306 on the container
+      - '3306'
+    # For Persistant data
+    volumes:
+      - local-db:/var/lib/mysql
+# Names our volume
+volumes:
+  local-db:
